groups

+++ src/main/java/com/vertex/vertex/group/service/GroupService.java
@@ -1,14 +1,11 @@
 package com.vertex.vertex.group.service;
 
-import com.vertex.vertex.comment.model.Comment;
 import com.vertex.vertex.group.model.dto.GroupDTO;
+import com.vertex.vertex.group.model.dto.GroupEditionDTO;
 import com.vertex.vertex.group.model.entity.Group;
 import com.vertex.vertex.group.repository.GroupRepository;
 import com.vertex.vertex.team.model.entity.Team;
 import com.vertex.vertex.team.service.TeamService;
-import com.vertex.vertex.user.model.entity.User;
-import com.vertex.vertex.user_team.model.dto.UserTeamDTO;
-import com.vertex.vertex.user_team.model.entity.UserTeam;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.BeanUtils;
@@ -31,16 +28,33 @@ public class GroupService {
         throw new EntityNotFoundException();
     }
 
+    public Group save(Long teamId, GroupEditionDTO dto) {
+        if (teamService.existsById(teamId)
+                && groupRepository.existsById(dto.getId())) {
+            Group gr = copyProps(dto, teamId);
+            return groupRepository.save(gr);
+        }
+        throw new EntityNotFoundException();
+    }
+
     public List<Group> findAll(){
         return groupRepository.findAll();
     }
 
     public Group findById(Long id){
-        return groupRepository.findById(id).get();
+        if (groupRepository.existsById(id)) {
+            return groupRepository.findById(id).get();
+        }
+        throw new EntityNotFoundException();
     }
 
-    public void deleteById(Long id){
-        groupRepository.deleteById(id);
+    public void deleteById(Long id) {
+        if (groupRepository.existsById(id)) {
+            groupRepository.deleteById(id);
+
+        } else {
+            throw new EntityNotFoundException();
+        }
     }
 
     public Group copyProps(
